Entidade y repo lito
aqui tan las entidades a su totalidad, creo, mas el repositorio, pero se debe añadir su logica que falto porque me sentia como el oyo, saludos

+++ gerente/src/main/java/com/example/gerente/model/Proveedor.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 public class Proveedor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_proveedor;
+    private Long id_proveedor;
     @Column(name = "nombre",nullable = false)
     private String nombre;
     @Column(name = "contacto", nullable = false)
